Move Configuration types outside their parent type
