+++ WordPress/Classes/ViewRelated/Support/SupportTableViewController.swift
@@ -87,7 +87,6 @@ class SupportTableViewController: UITableViewController {
         dismissTapped?()
         dismiss(animated: true)
     }
-
 }
 
 // MARK: - Private Extension
